requirements.txt
Flask==3.0.3
PyMuPDF==1.24.9
Pillow==10.4.0
